requirements.txt
apispec[yaml]
apispec-webframeworks
cryptography
dependencies
flask
flask-cors
flask-restful
flask-migrate
flask-sqlalchemy
flask-marshmallow
flask-jwt-extended
Flask-UUID
google-auth
google_auth_oauthlib
gunicorn
marshmallow
marshmallow-sqlalchemy
passlib
python-dotenv
psycopg2-binary
returns
stories
tox
webargs
